Error handling

+++ module_09/ex00/BitcoinExchange.hpp
@@ -15,11 +15,11 @@ class	BitcoinExchange {
 
 	public:
 
-		std::multimap<time_t, int> _data;
-		std::multimap<time_t, int> _values;
+		std::multimap<time_t, float> _data;
+		std::multimap<time_t, float> _values;
 
 		BitcoinExchange 			( void );
-		BitcoinExchange 			( std::multimap<time_t, int> data, std::multimap<time_t, int> values );
+		BitcoinExchange 			( std::multimap<time_t, float> data, std::multimap<time_t, float> values );
 		BitcoinExchange 			( const BitcoinExchange & var );
 		~BitcoinExchange			( void );
 		BitcoinExchange &operator=	(const BitcoinExchange &tmp);
@@ -27,7 +27,7 @@ class	BitcoinExchange {
 		void	get_values( void );
 };
 std::fstream										&read_file			(std::string file);
-std::multimap<time_t, int>	load_values			(std::fstream &values, char separator);
+std::multimap<time_t, float>	load_values			(std::fstream &values, char separator);
 std::ostream &operator<<(std::ostream& os, const BitcoinExchange &tmp);
 
 #endif
